Refactor HTTP module into broker components
• Rename server.rs to broker/ directory structure
• Split route handlers into separate modules: producer, consumer, metadata
• Move request/response types from broker to common.rs
• Add centralized routes configuration
• Remove duplicate poll_records_for_consumer_group function
• Add get_topics endpoint with TopicsResponse type
• Fix module inception warning by renaming broker.rs to server.rs
