+++ src/http/broker/mod.rs
@@ -1,0 +1,7 @@
+pub mod consumer;
+pub mod metadata;
+pub mod producer;
+pub mod routes;
+pub mod server;
+
+pub use server::start_server as start_broker;
